refactor: apply clippy lints and improve code quality
- Replace manual string prefix checks with strip_prefix()
- Use idiomatic Rust patterns (is_empty(), div_ceil(), iterator methods)
- Simplify error handling with direct ApiError::from conversions
- Fix function naming conventions (PascalCase to snake_case)
- Remove unnecessary borrows, type parameters, and imports
- Add type aliases for complex return types
- Move test-only helper functions into test modules
- Use more efficient iterator patterns instead of manual loops

No functional changes - purely code quality improvements.

+++ crates/sign/src/pq/dilithium/polyvec.rs
@@ -205,7 +205,7 @@ pub fn expand_matrix_a<P: DilithiumSchemeParams>(
                 xof.squeeze(&mut temp_buf).map_err(SignError::from_algo)?;
                 // Extract two 12-bit values from 3 bytes
                 let d1 = (temp_buf[0] as u32) | (((temp_buf[1] as u32) & 0x0F) << 8);
-                let d2 = (((temp_buf[1] as u32) >> 4) | ((temp_buf[2] as u32) << 4)) as u32;
+                let d2 = ((temp_buf[1] as u32) >> 4) | ((temp_buf[2] as u32) << 4);
 
                 if d1 < DilithiumParams::Q {
                     poly.coeffs[ctr] = d1;
